testing warning message for lack of api key when self revoking and updating proxied authenticator check plus test
