+++ tests/test_authentication.py
@@ -5,12 +5,9 @@ import shutil
 import subprocess
 import sys
 import time
-from typing import Any, Dict, Union
 
-import httpx
 import numpy
 import pytest
-from respx import MockRouter
 from starlette.status import (
     HTTP_400_BAD_REQUEST,
     HTTP_401_UNAUTHORIZED,
@@ -61,110 +58,6 @@ def config(sqlite_or_postgres_uri):
             },
         ],
     }
-
-
-@pytest.fixture
-def well_known_url() -> str:
-    return "http://example.com/well_known/"
-
-
-@pytest.fixture
-def tokens_response() -> Dict[str, str]:
-    return {
-        "access_token": "jwt",
-        "expires_in": "60",
-        "refresh_expires_in": "0",
-        "refresh_token": "jwt",
-        "token_type": "Bearer",
-        "id_token": "jwt",
-        "not-before-policy": "0",
-        "session_state": "uuid",
-        "scope": "offline_access email",
-    }
-
-
-@pytest.fixture
-def oidc_config(well_known_url: str) -> Dict[str, Any]:
-    return {
-        "authentication": {
-            "secret_keys": ["SECRET"],
-            "providers": [
-                {
-                    "provider": "keycloak_oidc",
-                    "authenticator": "tiled.authenticators:ProxiedOIDCAuthenticator",
-                    "args": {
-                        "audience": "tiled_aud",
-                        "client_id": "tiled",
-                        "device_flow_client_id": "tiled-cli",
-                        "well_known_uri": well_known_url,
-                        "confirmation_message": "You have logged in with Proxied OIDC as {id}.",
-                        "scopes": ["create:apikeys"],
-                    },
-                }
-            ],
-        },
-        "trees": [
-            {
-                "tree": f"{__name__}:tree",
-                "path": "/",
-            },
-        ],
-    }
-
-
-@pytest.fixture
-def mock_oidc_server(
-    respx_mock: MockRouter,
-    base_url: str,
-    well_known_url: str,
-    well_known_response: Dict[str, Any],
-    tokens_response: Dict[str, Union[str, int]],
-    oidc_config: Dict[str, Any],
-) -> MockRouter:
-    respx_mock.get(well_known_url).mock(
-        return_value=httpx.Response(httpx.codes.OK, json=well_known_response)
-    )
-
-    device_flow_client_id = oidc_config["authentication"]["providers"][0]["args"][
-        "device_flow_client_id"
-    ]
-    device_code = "FsYWEv-Fl4wkFlrtp-EWH7HR1pkCG2NIfBNeUKlZBAY"
-    user_code = "LCWE-ROXW"
-    verification_uri = f"{base_url}device"
-    verification_uri_complete = f"{base_url}device?user_code={user_code}"
-
-    respx_mock.post(
-        well_known_response["device_authorization_endpoint"],
-        data={"client_id": device_flow_client_id, "scope": "offline_access openid"},
-    ).mock(
-        return_value=httpx.Response(
-            status_code=httpx.codes.OK,
-            json={
-                "device_code": device_code,
-                "user_code": user_code,
-                "verification_uri": verification_uri,
-                "verification_uri_complete": verification_uri_complete,
-                "expires_in": 600,
-                "interval": 5,
-            },
-        )
-    )
-
-    respx_mock.post(
-        well_known_response["token_endpoint"],
-        data={
-            "device_code": device_code,
-            "grant_type": "urn:ietf:params:oauth:grant-type:device_code",
-            "client_id": device_flow_client_id,
-        },
-        name="token_polling",
-    ).mock(
-        return_value=httpx.Response(
-            status_code=httpx.codes.OK,
-            json=tokens_response,
-        )
-    )
-    return respx_mock
 
 
 def test_password_auth(enter_username_password, config):
@@ -824,9 +717,7 @@ def test_api_key_self_revoked(enter_username_password, config):
             api_context.whoami()
 
 
-def test_unallowed_api_key_self_revoked(
-    enter_username_password, config, oidc_config, mock_oidc_server
-):
+def test_unallowed_api_key_self_revoked(enter_username_password, config):
     with Context.from_app(build_app_from_config(config)) as context:
         # this logs in using username and password
         with enter_username_password("alice", "secret1"):
@@ -847,12 +738,6 @@ def test_unallowed_api_key_self_revoked(
 
     with Context.from_app(
         build_app_from_config(config), api_key=third_api_key["secret"]
-    ) as api_context:
-        with fail_with_status_code(HTTP_401_UNAUTHORIZED):
-            api_context.revoke_api_key(third_api_key["first_eight"])
-
-    with Context.from_app(
-        build_app_from_config(oidc_config), api_key=third_api_key["secret"]
     ) as api_context:
         with fail_with_status_code(HTTP_401_UNAUTHORIZED):
             api_context.revoke_api_key(third_api_key["first_eight"])
@@ -902,6 +787,16 @@ def test_admin_can_self_revoke(enter_username_password, config):
             build_app_from_config(config), api_key=api_key["secret"]
         ) as api_context:
             api_context.whoami()
+
+
+def test_need_api_key_to_self_revoke(enter_username_password, config):
+    with Context.from_app(build_app_from_config(config)) as context:
+        with enter_username_password("alice", "secret1"):
+            context.authenticate()
+            with pytest.warns(
+                match="You must be authenticated with an API key to revoke your API key."
+            ):
+                context.revoke_self_api_key()
 
 
 @pytest.mark.timeout(5)
